+++ tables/specialClothing/index.php
@@ -44,7 +44,7 @@ if (mysqli_num_rows($result) > 0) {
                     </a>
                 </td>
                 <td>
-                    <a class='iconButton' href='delete.php?id=" . $row['id'] . "'>
+                    <a class='iconButton' href='delete.php?id=" . $row['id'] . "' onclick='return confirmDelete();'>
                         <img src='/public/icon/delete-icon.svg' width='20' height='20' alt='удалить'>
                     </a>
                 </td>        
@@ -58,5 +58,10 @@ mysqli_close($mysqli);
 ?>
 
 </section>
+<script>
+    function confirmDelete() {
+        return confirm("Вы уверены, что хотите удалить эту запись?");
+    }
+</script>
 </body>
 </html>
